[feat] Sistema de recuperação Editar

+++ Flask-app/routes/deliverys_routes.py
@@ -175,7 +175,7 @@ def EditDeviveryDatas():
         name = formsData['name']
         clientId = formsData['clientId']
         idType = formsData['typeClientId']
-        
+        ProductController().RestoreProductsQuantityFromADeliveryId(datas['company_email'],DeliveryController().GetProductsFromDelivery(datas['company_email'],delivery_id))
         ok = DeliveryController().EditDelivery(datas['company_email'],delivery_id,items,address,date,name,clientId,idType)
         if ok:
             return jsonify({'status':'ok'}),200
@@ -199,9 +199,11 @@ def DeleteDelivery():
         delivery_id= request.get_json()['delivery_id']
         ProductController().RestoreProductsQuantityFromADeliveryId(datas['company_email'],DeliveryController().GetProductsFromDelivery(datas['company_email'],delivery_id))
         delivery_was_deleted = DeliveryController().DeleteDelivery(datas['company_email'],delivery_id)
-        products_deleted = DeliveryController().DeleteProduct(datas['company_email'],delivery_id)
-        if products_deleted:
-            return jsonify({'status':'ok'}),200
+        
+        if delivery_was_deleted:
+            products_deleted = DeliveryController().DeleteProduct(datas['company_email'],delivery_id)
+            if products_deleted:
+                return jsonify({'status':'ok'}),200
 
         return jsonify({'status':'error'}),500
